Move theme logic from app.js to theme.js
